fix: reset envs when user switches workspaces (#3039)
